@@ -30,18 +30,17 @@
 
 package com.raywenderlich.android.redditclone
 
-import android.support.v7.util.DiffUtil
+
+import androidx.recyclerview.widget.DiffUtil
 import com.raywenderlich.android.redditclone.networking.RedditPost
 
 class RedditDiffUtilCallback : DiffUtil.ItemCallback<RedditPost>() {
+  override fun areItemsTheSame(oldItem: RedditPost, newItem: RedditPost): Boolean {
+    return oldItem?.key == newItem?.key  }
 
-  override fun areItemsTheSame(oldItem: RedditPost?, newItem: RedditPost?): Boolean {
-    return oldItem?.key == newItem?.key
-  }
+  override fun areContentsTheSame(oldItem: RedditPost, newItem: RedditPost): Boolean {
+    return oldItem?.title == newItem?.title
+            && oldItem?.score == newItem?.score
+            && oldItem?.commentCount == newItem?.commentCount  }
 
-  override fun areContentsTheSame(oldItem: RedditPost?, newItem: RedditPost?): Boolean {
-  return oldItem?.title == newItem?.title
-      && oldItem?.score == newItem?.score
-      && oldItem?.commentCount == newItem?.commentCount
-  }
 }
