@@ -30,10 +30,10 @@
 
 package com.raywenderlich.android.redditclone.database
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
 import android.content.Context
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
 import com.raywenderlich.android.redditclone.networking.RedditPost
 
 @Database(
