@@ -30,8 +30,9 @@
 
 package com.raywenderlich.android.redditclone.networking
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
 
 @Entity
